requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
iniconfig==2.0.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
Pygments==2.18.0
pytest==8.2.1
python-decouple==3.8
sqlparse==0.5.0
